refactor: fetchData 함수명 변경에 따른 import 경로 수정
